pyproject.toml
[tool.poetry]
name = "iracingsetups_client"
version = "0.1.0"
description = ""
authors = ["Leon Weemen <weemen@leonweemen.nl>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
uvloop = "^0.19.0"
configloader = "^1.0.1"
pyyaml = "^6.0.1"
pyirsdk = "^1.3.5"

[tool.poetry.scripts]
app = "iracingsetups_client.app:main"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
